requirements.txt
aiohttp~=3.7.4.post0
aiosmtpd~=1.4.2
APScheduler~=3.7.0
async-timeout~=3.0.1
atpublic~=2.1.3
attrs~=20.3.0
chardet~=4.0.0
confuse~=1.4.0
idna~=3.1
msgraph-async~=0.1.1
multidict~=5.1.0
python-dotenv~=0.17.0
pytz~=2021.1
PyYAML~=5.4.1
six~=1.15.0
typing-extensions~=3.7.4.3
tzlocal~=2.1
yarl~=1.6.3
argon2-cffi~=20.1.0